Update README.md
Young people can prepare for a career in IT by reading "A Peek into Java for Nigerian Teenagers." But how do we ensure that they are not discouraged by big words like "Class" and unfamiliar phrases like "case sensitive." Our objective here is to teach young people the difficult words and phrases that prevent them from reading the book from cover to cover. To help them get through the book, we try to explain the tough words in the book, one word at a time.